refactor(splunk): split the form part to component
EVNT-341
